requirement.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.2
djangorestframework==3.14.0
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.2
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
translate==3.6.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
